snp-analyzer/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
xlrd==2.0.1
openpyxl==3.1.5
pandas==2.2.3
numpy==2.2.1
scikit-learn
matplotlib
reportlab
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
